requirements.txt
ultralytics==8.0.145
tkinter==

opencv-python~=4.8.0.74
Pillow~=10.0.0